// src/models/candidatos.ts
export type Candidatos = {
    id: number; // Adicione o campo id se necessário
    user_id: number;
    ultimo_cargo: string;
    ultimo_salario: number;
    pretensao_salarial_clt: number;
    pretensao_salarial_pj: number;
    posicao_desejada: string;
    escolaridade: string;
    graduacao_principal: string;
    como_conheceu: string;
    consultor_talents: string;
    nivel_ingles: string;
    qualificacoes_tecnicas: string;
    certificacoes: string;
    cv: string;
    created_at: string;
}